an ARP responder for routers/firewalls on community ports.
this should allow clients that are isolated from each other to
communicate via the router or firewall on the subnet. this is
acheived by having the router reply to arp requests by these clients
with it's own l2 address, which in turn directs all l3 communication
to the router for... routing.

the idea came from alex wilson. i had the time and a lot of code i
could repurpose for this.