protect the tunnelled arp packet
this uses chacha20-poly1305, which is an extremely heavy hammer for
this little nail, but it's nice to get some experience with it
before trying to apply it in bigger settings.

the biggest problem with this is that it doesn't prevent replays,
but the original code didnt either. it's still a step forward and
no step back.